seperated combining of mesh into own class
