+++ 3DAmsterdam/Assets/3DAmsterdam/Scenes/DataGeneration/GenerateBuildingTilesFromOBJ.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aed2c26805976c148b59dfb54ed635e7
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
