+++ 3DAmsterdam/Assets/3DAmsterdam/Scenes/DataGeneration/GenerateBuildingTilesFromOBJSettings.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d833aef3fc412e94a8369a4c761687e1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
