+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/AssetGeneration/TileCombiner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41c7cd272ef313c4480c0f2a77d5cb57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
