+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/ModelParsing/WeldMeshVertices.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf9c49c4c20bc734484ef55f3842b0d4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
