+++ 3DAmsterdam/UserSettings/EditorUserSettings.asset
@@ -5,6 +5,15 @@ EditorUserSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 4
   m_ConfigSettings:
+    RecentlyUsedScenePath-0:
+      value: 2242470311464608292f013104264d14171a57192f2d28353e663932f6e1133dece224e8f33330327625ea32012c002bf7281c02fa0f0305163af40214ec5e061fcc0cdc
+      flags: 0
+    RecentlyUsedScenePath-1:
+      value: 2242470311464608292f013104264d14171a57192f2d28353e663932f6e1133dece224e8f33330327625ea32012c002bf7281c02fa0f0305163af40214ec36011ec83fc015e22c33e55709c0c409c8
+      flags: 0
+    RecentlyUsedScenePath-2:
+      value: 2242470311464608292f013104264d14171a57192f2d28353e663932f6e1133dece224e8f33330327625ea32012c002bf7281c02fa0f0305163af40214ec36011ec837e730860a12c20d05
+      flags: 0
     UnityEditor.ShaderGraph.FloatingWindowsLayout2:
       value: 181344140043005e1a220d3b1f364b524c0c5a27130c293326201334cee5322ca0bd30e8eb293a707b0fd0180b3d0a36fc0d3d04e649500d1002ee0b5dbd1d2c27c00ad113cb1e10e41f1addc80993b9859884a69ae6d8f0d1cda9e8fbfefaf9f9dea3fdb9ade882f0ffb0e1e380cafbf2c3adc18e9cd285a2908b87ee819c8395949c9483d38b8d97ddbd90eed2a5a892a2af949aa48bafb19f85bd75a7ed3a7d25658598b7b58bb4b76aaf777690ca2863946c72c6cd81b6b6708f9f879f88769589d91c8f888e64d20f935e796571755c6b546677696a486c781c073c6f23696429450074652134203e56454c514e4352305f78621d07746f0243326d645d4b5f450e471441155071767740736e6130644f727d47667b34007c5f3fe12420e62304f4e023e03efadde518f54165ebb4efba505858565e5d45bf0a809a4048928946d4da
       flags: 0
@@ -18,5 +27,7 @@ EditorUserSettings:
   m_SemanticMergeMode: 2
   m_VCShowFailedCheckout: 1
   m_VCOverwriteFailedCheckoutAssets: 1
-  m_VCOverlayIcons: 1
+  m_VCProjectOverlayIcons: 1
+  m_VCHierarchyOverlayIcons: 1
+  m_VCOtherOverlayIcons: 1
   m_VCAllowAsyncUpdate: 0
